+++ src/Application/AnalyticsAppService.cs
@@ -1,0 +1,7 @@
+namespace ViajaNet.JobApplication.Application
+{
+    public class AnalyticsAppService
+    {
+
+    }
+}