+++ src/Infrastructure/Queue/QueueFactory.cs
@@ -1,0 +1,8 @@
+namespace ViajaNet.JobApplication.Infrastructure.Queue
+{
+    public class QueueFactory
+    {
+        public QueueFactory()
+        { }
+    }
+}